form validation in user register set, user search in admin side: get result by searching email or user name- done,sweet alert and sweet conformation added in admin side

+++ frontend/src/components/UsersData.js
@@ -2,6 +2,8 @@ import { Button, Spinner } from "react-bootstrap";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteUser, blockUser } from '../features/admin/usersSlice'
 import { useEffect } from "react";
+import Swal from 'sweetalert2'
+
 
 
 function UsersData({ data, index }) {
@@ -9,12 +11,47 @@ function UsersData({ data, index }) {
     const dispatch = useDispatch()
 
     const delUser = (userId) => {
-        dispatch(deleteUser(userId))
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+                confirmButton: 'btn btn-success',
+                cancelButton: 'btn btn-danger'
+            },
+            buttonsStyling: false
+        })
+
+        swalWithBootstrapButtons.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteUser(userId))
+
+                swalWithBootstrapButtons.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            } else if (
+                /* Read more about handling dismissals below */
+                result.dismiss === Swal.DismissReason.cancel
+            ) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelled',
+                    'Your imaginary file is safe :)',
+                    'error'
+                )
+            }
+        })
     }
 
     return (
-       
-         <tr>
+
+        <tr>
             <td>{index + 1}</td>
             <td>{data.name}</td>
             <td>{data.email}</td>
@@ -35,16 +72,16 @@ function UsersData({ data, index }) {
                 )}{" "}
             </td>
             <td>
-                    <Button
-                        onClick={() => delUser(data._id)}
-                        style={{ backgroundColor: "red" }}
-                    >
-                        {" "}
-                        Delete
-                    </Button>
+                <Button
+                    onClick={() => delUser(data._id)}
+                    style={{ backgroundColor: "red" }}
+                >
+                    {" "}
+                    Delete
+                </Button>
             </td>
         </tr>
-    
+
     );
 }
